+++ pylib/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2023-present Oori Data <info@oori.dev>
 # SPDX-License-Identifier: Apache-2.0
 # onya.__about__
-__version__ = '0.4.0'
+__version__ = '0.4.1'